+++ src/main/java/com/khomaytinh/quanlykhomaytinh/Model/Mapper/TaiNgheMapper.java
@@ -14,7 +14,7 @@ public class TaiNgheMapper implements RowMapper<TaiNghe> {
         hh.setID(rs.getString("MAHH"));
         hh.setTen(rs.getString("Ten"));
         hh.setHangSX(rs.getString("HangSX"));
-        hh.setGia(rs.getDouble("Gia"));
+        hh.setGia((long)rs.getDouble("Gia"));
         hh.setHinhAnh(rs.getBlob("HinhAnh"));
         hh.setTrongLuong(rs.getDouble("TrongLuong"));
         hh.setSoLuongCon(rs.getInt("SoLuongCon"));
